+++ sdks/outpost-typescript/src/funcs/tenantsUpsert.ts
@@ -35,7 +35,7 @@ import { Result } from "../types/fp.js";
 export function tenantsUpsert(
   client: OutpostCore,
   tenantId: string,
-  params?: components.TenantUpsert | undefined,
+  body?: components.TenantUpsert | undefined,
   options?: RequestOptions,
 ): APIPromise<
   Result<
@@ -56,7 +56,7 @@ export function tenantsUpsert(
   return new APIPromise($do(
     client,
     tenantId,
-    params,
+    body,
     options,
   ));
 }
@@ -64,7 +64,7 @@ export function tenantsUpsert(
 async function $do(
   client: OutpostCore,
   tenantId: string,
-  params?: components.TenantUpsert | undefined,
+  body?: components.TenantUpsert | undefined,
   options?: RequestOptions,
 ): Promise<
   [
@@ -87,7 +87,7 @@ async function $do(
 > {
   const input: operations.UpsertTenantRequest = {
     tenantId: tenantId,
-    params: params,
+    body: body,
   };
 
   const parsed = safeParse(
@@ -99,7 +99,7 @@ async function $do(
     return [parsed, { status: "invalid" }];
   }
   const payload = parsed.value;
-  const body = encodeJSON("body", payload.params, { explode: true });
+  const body$ = encodeJSON("body", payload.body, { explode: true });
 
   const pathParams = {
     tenant_id: encodeSimple("tenant_id", payload.tenant_id, {
@@ -140,7 +140,7 @@ async function $do(
     baseURL: options?.serverURL,
     path: path,
     headers: headers,
-    body: body,
+    body: body$,
     userAgent: client._options.userAgent,
     timeoutMs: options?.timeoutMs || client._options.timeoutMs || -1,
   }, options);