+++ sdks/outpost-typescript/src/lib/config.ts
@@ -65,8 +65,8 @@ export function serverURLFromOptions(options: SDKOptions): URL | null {
 export const SDK_METADATA = {
   language: "typescript",
   openapiDocVersion: "0.0.1",
-  sdkVersion: "0.8.0",
-  genVersion: "2.849.10",
+  sdkVersion: "0.9.0",
+  genVersion: "2.862.0",
   userAgent:
-    "speakeasy-sdk/typescript 0.8.0 2.849.10 0.0.1 @hookdeck/outpost-sdk",
+    "speakeasy-sdk/typescript 0.9.0 2.862.0 0.0.1 @hookdeck/outpost-sdk",
 } as const;